+++ changelog.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- Modified page [/create-polygroup-from-mask-in-zbrush](https://codyburleson.com/create-polygroup-from-mask-in-zbrush); minor content edits to improve quality / readability.
+
 ## [1.5.16] - 2020-01-30]
 
 - Added page [/book-review-the-goldfinch](https://codyburleson.com/book-review-the-goldfinch)